@@ -1,14 +1,14 @@
 name: example
 
-environment:  
-  sdk: ">=2.7.2 <3.0.0"
+environment:
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
-  copy_with_e_annotation: ^1.0.1
+  copy_with_e_annotation: ^1.0.2-nullsafety2
 
 dev_dependencies:
   build_runner: ^1.7.3
-  test: ^1.9.3
+  test: ^1.16.0-nullsafety.16
   meta: ^1.1.8
   copy_with_e_generator:
     path: ../copy_with_e_generator/
