@@ -1,5 +1,5 @@
 class CopyWithE {
-  final List<Type> types;
+  final List<Type>? types;
 
   const CopyWithE([this.types]);
 }
