@@ -1,11 +1,12 @@
 import 'package:copy_with_e_annotation/copy_with_e_annotation.dart';
-import 'package:meta/meta.dart';
 import 'package:test/test.dart';
 
 part 'ex4_test.g.dart';
 
 main() {
-  test("ALL FIELDS FROM MY CONCRETE EXTENDED SUPERCLASS SHOULD BE INCLUDED IN MY COPYWITH EXTENSION METHOD", () {
+  test(
+      "ALL FIELDS FROM MY CONCRETE EXTENDED SUPERCLASS SHOULD BE INCLUDED IN MY COPYWITH EXTENSION METHOD",
+      () {
     var rob = Employee(age: 9, name: "rob", id: 5);
     var robEdited = rob.cwPerson(age: 3);
 
@@ -24,8 +25,8 @@ class Person implements HasAge {
   final String name;
 
   Person({
-    @required this.age,
-    @required this.name,
+    required this.age,
+    required this.name,
   });
 }
 
@@ -34,9 +35,9 @@ class Employee extends Person {
   final int id;
 
   Employee({
-    @required this.id,
-    @required age,
-    @required name,
+    required this.id,
+    required age,
+    required name,
   }) : super(
           age: age,
           name: name,
