@@ -29,10 +29,14 @@ main() {
             name: "Bad " + petOwner.name,
             pets: () {
               if (petOwner is DogOwner) //
-                return petOwner.pets.map<Dog>((x) => x.cwPet(isAlive: false)).toList();
+                return petOwner.pets
+                    .map<Dog>((x) => x.cwPet(isAlive: false) as Dog)
+                    .toList();
 
               if (petOwner is CatOwner) //
-                return petOwner.pets.map<Cat>((x) => x.cwPet(isAlive: false)).toList();
+                return petOwner.pets
+                    .map<Cat>((x) => x.cwPet(isAlive: false) as Cat)
+                    .toList();
 
               throw Exception("unexpected owner");
             }(),
@@ -58,18 +62,26 @@ class DogOwner implements PetOwnerBase<Dog, int> {
   final int id;
   final List<Dog> pets;
   final String name;
-  final String dogStuff;
+  final String? dogStuff;
 
-  DogOwner({this.id, this.pets, this.name, this.dogStuff});
+  DogOwner(
+      {required this.id,
+      required this.pets,
+      required this.name,
+      this.dogStuff});
 }
 
 class CatOwner implements PetOwnerBase<Cat, String> {
   final String id;
   final List<Cat> pets;
   final String name;
-  final String catStuff;
+  final String? catStuff;
 
-  CatOwner({this.id, this.pets, this.name, this.catStuff});
+  CatOwner(
+      {required this.id,
+      required this.pets,
+      required this.name,
+      this.catStuff});
 }
 
 @CopyWithE()
@@ -81,12 +93,12 @@ class Dog extends Pet {
   final bool isAlive;
   final String woofType;
 
-  Dog({this.isAlive, this.woofType});
+  Dog({required this.isAlive, required this.woofType});
 }
 
 class Cat extends Pet {
   final bool isAlive;
   final String whiskers;
 
-  Cat({this.isAlive, this.whiskers});
+  Cat({required this.isAlive, required this.whiskers});
 }
