@@ -11,7 +11,8 @@ import 'package:source_gen/source_gen.dart';
 
 import 'GeneratorForAnnotationX.dart';
 
-List<NameType> getAllFields(List<InterfaceType> interfaceTypes, ClassElement element) {
+List<NameType> getAllFields(
+    List<InterfaceType> interfaceTypes, ClassElement element) {
   var superTypeFields = interfaceTypes //
       .where((x) => x.element.name != "Object")
       .flatMap((st) => st.element.fields.map((f) => //
@@ -33,9 +34,10 @@ class CopyWithEGenerator extends GeneratorForAnnotationX<CopyWithE> {
   ) {
     var sb = StringBuffer();
 
-    sb.writeln("//RULES: 1 all subtypes must be in same file or be passed in. 2 the types passed in must all be classes");
+    sb.writeln(
+        "//RULES: 1 all subtypes must be in same file or be passed in. 2 the types passed in must all be classes");
 
-    var types = List<ClassDef>();
+    var types = <ClassDef>[];
     if (!annotation.read('types').isNull) {
       types = annotation
           .read('types') //
@@ -44,7 +46,8 @@ class CopyWithEGenerator extends GeneratorForAnnotationX<CopyWithE> {
         var el = x.toTypeValue().element;
 
         if (el is! ClassElement) {
-          throw Exception("the list of types for the copywith def must all be classes");
+          throw Exception(
+              "the list of types for the copywith def must all be classes");
         }
 
         var ce = (el as ClassElement);
@@ -54,14 +57,18 @@ class CopyWithEGenerator extends GeneratorForAnnotationX<CopyWithE> {
           ce.name,
           getAllFields(ce.allSupertypes, ce),
           [],
-          [...ce.interfaces.map((e) => e.element.name), ce.supertype.element.name],
+          [
+            ...ce.interfaces.map((e) => e.element.name),
+            ce.supertype.element.name
+          ],
         );
       }).toList();
     }
 
     if (element is ClassElement) {
       var subClasses = allClasses //
-          .where((x) => x.allSupertypes.any((st) => st.element.name == element.name))
+          .where((x) =>
+              x.allSupertypes.any((st) => st.element.name == element.name))
           .where((x) => !types.any((t) => t.name == x.name))
           .map((x) => ClassDef(
                 x.isAbstract,
@@ -78,17 +85,28 @@ class CopyWithEGenerator extends GeneratorForAnnotationX<CopyWithE> {
         element.isAbstract,
         element.name,
         getAllFields(element.allSupertypes, element),
-        element.typeParameters.isEmpty ? [] : element.typeParameters.where((x) => x.name != null).map((x) => //
-            GenericType(x.name, x.bound == null ? null : x.bound.toString())).toList(),
-        [...element.interfaces.map((e) => e.element.name), element.supertype.element.name],
+        element.typeParameters.isEmpty
+            ? []
+            : element.typeParameters
+                .where((x) => x.name != null)
+                .map((x) => //
+                    GenericType(
+                        x.name, x.bound == null ? null : x.bound.toString()))
+                .toList(),
+        [
+          ...element.interfaces.map((e) => e.element.name),
+          element.supertype.element.name
+        ],
       );
 
-      sb.writeln("//" + types2.map((e) => e.name).toList().toString());
+      sb.writeln("/*" + extClass.fields.toString() + "*/");
 
       sb.writeln(createCopyWith(extClass, types2));
     }
 
-    return element.session.getResolvedLibraryByElement(element.library).then((resolvedLibrary) {
+    return element.session
+        .getResolvedLibraryByElement(element.library)
+        .then((resolvedLibrary) {
       return sb.toString();
     });
   }
