@@ -23,12 +23,16 @@ int getDepth(int count, ClassDef thisType, List<ClassDef> types) {
   var related = thisType.baseTypes.intersect(types.map((e) => e.name)).toList();
 
   if (related.length > 1) //
-    return related.map((r) => getDepth(count + 1, types.firstWhere((x) => x.name == r), types)).max();
+    return related
+        .map((r) =>
+            getDepth(count + 1, types.firstWhere((x) => x.name == r), types))
+        .max();
 
   if (related.length == 0) //
     return count;
 
-  return getDepth(count + 1, types.firstWhere((x) => x.name == related[0]), types);
+  return getDepth(
+      count + 1, types.firstWhere((x) => x.name == related[0]), types);
 }
 
 String getCopyWithParamList(
@@ -36,7 +40,7 @@ String getCopyWithParamList(
   List<GenericType> generics,
 ) {
   return fields
-      .map((e) => "${e.type} ${e.name}") //
+      .map((e) => "${e.type.replaceAll("?", "")}? ${e.name}") //
       .joinToString(separator: ", ");
 }
 
@@ -68,18 +72,18 @@ String getGenericParams(List<GenericType> generics) {
   return "<$genericList>";
 }
 
-String getPropertySetThis(String className, String fieldName) => //
-    "$fieldName: (this as $className).$fieldName";
+String getPropertySetThis(String className, String fieldName, String type) => //
+    "$fieldName: (this as $className).$fieldName as $type";
 
-String getPropertySet(String name) => //
-    "$name: $name == null ? this.$name : $name";
+String getPropertySet(String name, String type) => //
+    "$name: $name == null ? this.$name as $type : $name as $type";
 
 String getConstructorLines(ClassDef extType, ClassDef typeType) {
   var result = typeType.fields.map((field) {
     if (extType.fields.any((x) => field.name == x.name)) {
-      return getPropertySet(field.name);
+      return getPropertySet(field.name, field.type);
     } else {
-      return getPropertySetThis(typeType.name, field.name);
+      return getPropertySetThis(typeType.name, field.name, field.type);
     }
   }).joinToString(separator: ",\n");
 
@@ -90,5 +94,7 @@ String getExtensionDef(String className) => //
     "extension ${className}Ext_CopyWithE on ${className}";
 
 String getConstructorName(String trimmedClassName) {
-  return trimmedClassName[trimmedClassName.length - 1] == "_" ? "$trimmedClassName._" : trimmedClassName;
+  return trimmedClassName[trimmedClassName.length - 1] == "_"
+      ? "$trimmedClassName._"
+      : trimmedClassName;
 }
