@@ -1,9 +1,8 @@
 import 'ex5_test.dart';
 
-class Stuff implements Person {
+class Stuff implements HasAge2 {
   final int age;
   final String name;
-  final String moreStuff;
 
-  Stuff({this.age, this.name, this.moreStuff});
+  Stuff({required this.age, required this.name});
 }
